requirements.txt
docker
pyyaml
docker-compose